Use Context rather than APIGatewayEventRequestContext
Note: APIGatewayEvent is deprecated and APIGatewayProxyEvent should be preferred.

I also added the APIGatewayProxyResult type to the handler as I think this is
useful to have in the example - it will fail to compile if the returned
value is missing attributes.

APIGatewayEventRequestContext is available on event.requestContext, it's not the
second argument for the handler.